requirements.txt
pandas>=2.0
numpy>=1.24
lightgbm>=4.0
numba>=0.58
scikit-learn>=1.3
cython>=3.0
pyyaml>=6.0
pyarrow>=14.0
holidays>=0.40
memory-profiler>=0.61
pyspark>=3.5
click>=8.1
pytest>=7.4
matplotlib>=3.7
